Move the queue trigger outside of services folder.
- Move the queue trigger file to `job` folder.
- Move the `job.js` inside `job` folder too.
- The queue trigger can be shared acrss different services.
- Update the entry script and build script.
