+++ src/services/azure/index.js
@@ -1,10 +1,8 @@
-import * as job from './job';
 import * as logger from './logger';
 import * as queue from './queue';
 import * as table from './table';
 
 export default {
-  job,
   logger,
   queue,
   table,
